Push documentation to GCS instead of Github pages.
www.bazel.io is now served from GCS and this change
archive the site on CI and adds a function to do the
push to GCS.

Another change to github pages is needed to redirect
http://bazel.io to http://www.bazel.io and a
change to ci.bazel.io to add the job that calls the
added method.

Moving towards #47.

--
MOS_MIGRATED_REVID=123126457