Add support for more augmented-assignment operators.
based on my limited understanding of python syntax, the only things we don't support is //= and **=, but it looks like skylark doesn't support the corresponding infix operators.

RELNOTES[NEW]: add support for the '-=', '*=', '/=', and'%=' operators to skylark.  Notably, we do not support '|=' because the semantics of skylark sets are sufficiently different from python sets.

--
MOS_MIGRATED_REVID=123889776