experimental UI: delay updates after action completion
Very often, when an action completes, a new one is started directly
afterwards. With this in mind, delay the refresh after the completion
of an action as long as acceptable by the refresh rate limiting. If
indeed an action starts within that time, we have replaced a double refresh
of the progress bar by a single one, thus make it appear less flickering.

--
Change-Id: I0e9143d6bba9929265ad2b6a3699486a1d5b3b2b
Reviewed-on: https://bazel-review.googlesource.com/#/c/3823
MOS_MIGRATED_REVID=125165516