Disable runfiles on Windows.
This adds a new configuration option that allows disabling the creation of symlink forest for runfiles.
On Windows, symlink forest is disabled by default; only the runfiles manifest is created.

For shell tests, a function 'rlocation' is provided that converts from runfiles location to a real location.

Work towards #1212.

--
MOS_MIGRATED_REVID=125439553