Rollback of commit b692917a73e73d05a65979992f87ad93445dee97.
*** Reason for rollback ***

Broke  Bazel build on Ubuntu 15.10, e.g.:
http://ci.bazel.io/view/Bazel%20bootstrap%20and%20maintenance/job/Bazel/JAVA_VERSION=1.8,PLATFORM_NAME=ubuntu_15.10-x86_64/611/console


*** Original change description ***


Make create_embedded_tools.sh an sh script

This script is called directly (and not as an argument to the shell),
hence the #!-line needs to be correct. As the script does not use any
bash-specific feature, we can maintain it as a script for the POSIX
shell, which has the advantage that the path is canonical.

--
MOS_MIGRATED_REVID=125441679