Replace "Getting Started" page by a Bazel "Hello World"
As we have a tutorial anyway, keep the "Getting started"
page short and language agnositic, and do not duplicate
parts that are taught in the tutorial anyway. The main
purpose of the "Getting Started" page is to give the
reader a quick feeling of the very basic Bazel concepts,
as well as allowing to test if Bazel is installed correctly.

--
Change-Id: I80baae1c424244d492c202eff0c613cbac187d43
Reviewed-on: https://bazel-review.googlesource.com/#/c/3882
MOS_MIGRATED_REVID=125796405