experimental UI: also during loading produce a short progress bar
When writing the experimental state tracker is asked to produce a progress
bar, it can be requested to have a one-line version. This one is used, e.g.,
if curses are not available and therefore it cannot be erased. Also honor
that request, if reporting about loading progress.

--
Change-Id: Ice3045fc86280e420b789a94b03427ff578f4a11
Reviewed-on: https://bazel-review.googlesource.com/#/c/3912
MOS_MIGRATED_REVID=126057511