Make experimental UI resistent to late-reported events
Sometimes the event reporting the completion of the last action
can, effectively, overtake the event reporting the completion of
last action. To avoid late writing to the output stream don't start
the update thread after the completion of the build is reported.
There is nothing to update then any more anyway.

--
Change-Id: Ibbb6d025635a9daa233c2b28ca64a1ae4896ce47
Reviewed-on: https://bazel-review.googlesource.com/#/c/3922
MOS_MIGRATED_REVID=126182576