Display all allowed types for SkylarkSignature Params that allow multiple types.
Fixes #921

Certain parameters, such as the `executable` and `command` parameters of
ctx.action, allow multiple types. However, the allowed types are not enumerated
in the SkylarkSignature annotation, causing the generated Skylark Library
document to not display any type information for those params.

This change adds a new field, `allowedTypes`, to `SkylarkSignature` that is a
list of `ParamType` objects. If the param can accept multiple types, then
`type` is set to `Object.class`, and `allowedTypes` is set to the list of types
that can be accepted.

--
MOS_MIGRATED_REVID=126617047