Add a final event on the event bus
Add an event to the event bus that is guaranteed to come after the
completion of the command. Listeners can use this event to clean up
any resources that still need to be cleaned up. As first customer, add
the ExperimentalEventHandler so that it is guaranteed not to leak the
update thread.

--
Change-Id: Ia20bae8a00541aa1338e49f74a7320ac098ab71d
Reviewed-on: https://bazel-review.googlesource.com/#/c/4011
MOS_MIGRATED_REVID=126909844