Create a symlink with the right workspace name under the execroot
The execution root currently uses the basename of the workspace directory for
the workspace name, not the name in the WORKSPACE file. (For example, if our
sources were in /path/to/foo and our WORKSPACE file had workspace(name = "bar"),
our execution root would look like execroot/foo.)

This creates a symlink bar -> foo, so that accessing ../repo_name actually works
for the main repository.

RELNOTES[INC]: The main repository's execution root is under the main
repository's workspace name, not the source directory's basename. This shouldn't
have any effect on most builds, but it's possible it could break someone doing
weird things with paths in actions.

--
MOS_MIGRATED_REVID=128175455