Sets SONAME on shared objects in Android binaries.
Adds a linker flag to set the internal DT_SONAME. This fixes #1578 for
SDK 24 and removes the warnings for previous SDKs. There is no need to
set the linker flag for android_librarys that depend on native code,
because the linker flag will be set by the android_binarys that depend
on that android_library.

--
Change-Id: If6422f2c2677ec974a3682e3038d7fe159ed4978
Reviewed-on: https://bazel-review.googlesource.com/#/c/4210/
MOS_MIGRATED_REVID=128833820