Continuing removing GUAVA dependencies from junitrunner.
Bazel users that are using a different Guava version than the one in the
junitrunner jar are getting an IncompatibleClassChangeError. This CL rewrites
parts of junitrunner code so it won't depend on Guava anymore.
Continuing progress on issue #1150. Removed Guava dependencies, except for
Supplier and Ticker, from junit.runner.junit4.

--
MOS_MIGRATED_REVID=129757029