Sets SONAME on shared objects in Android binaries.
Adds a linker flag to set the internal DT_SONAME. This fixes #1578 for
SDK 24 and removes the warnings for previous SDKs. There is no need to
set the linker flag for android_librarys that depend on native code,
because the linker flag will be set by the android_binarys that depend
on that android_library.

Testing is done via `readelf` executable that is shipped with the NDK.

--
Change-Id: I16fdfe6522c8694ce51554289122bf035a61d9ec
Reviewed-on: https://bazel-review.googlesource.com/#/c/4302/
MOS_MIGRATED_REVID=129920256