Rollback of commit fbaa700337bf0cf9083f083e6cd0d2f82faf1506.
*** Reason for rollback ***

I want to do this in a way that exposes the data attribute only for android_test and android_binary, so I'll add it to the deploy manifest instead.

*** Original change description ***

Expose test "data" attribute to IDE.

This is needed to deploy android_tests to the device for testing. For
symmetry we expose it for all tests rules.

--
MOS_MIGRATED_REVID=129926351