Support setting the build id via the client environment
Normally, Bazel generates a new UUID for every build invocation. In some
use cases, however, the invocation of Bazel is part of a larger build
process or otherwise controlled by a different tool. In this case, it
is useful to tell Bazel the identifier of the build to make it fit with
other identifiers generated by the controlling process. To achieve this,
the variable BAZEL_INTERNAL_INVOCATION_ID from the client environment
is inspected and if it is set and its value is a syntactically correct
UUID, this UUID will be used as the identifier for the build. It is in
the responsibility of the caller to ensure the id is sufficiently unique
if that environment variable is set.

--
MOS_MIGRATED_REVID=130079446