Handle public tags in merger -> R.class writer
Developers do use public tags, and before this change
it would assert in the AndroidResourceClassWriter
as an unhandled ResourceType. We probably didn't want
to write out a R.public.field anyway.

Also, handle public tags with the same name,
but different type. They get mapped to the same FQN,
so use the combining mechanism to keep track of
the different types and ids.

--
MOS_MIGRATED_REVID=130395089