Adds data binding support to the Android resource processor.
This is invoked via the new --dataBindingInfoOut flag.

If specified and the rule has resources, data binding declarations
in layout .xmls are stripped and fed into an aggregated .zip file
under that name. The original (stripped) .xml is then fed into
aapt for normal processing.

If specified and the rule has no resources, an empty file is created.

If not specified, everything works as before.

--
MOS_MIGRATED_REVID=130973265