Rollback of commit 0545788b86c8ff58f22ac62e51d096e876e7feea.
*** Reason for rollback ***

Breaks tachyon because of the fact that it has duplicate symbol errors but they get discarded when compiled into the same archive, which is not the case anymore as we are being more clever with dependencies....

*** Original change description ***

Instead of using the smallest group as the inputs for the generation actions, use instead the intersection of all the groups in which each proto appears. This further minimizes the average number of inputs per generation/compilation action.

--
MOS_MIGRATED_REVID=131099657