Cloned from commit 2dddbeeac25ace6f53dd3ce6ac588111c239bdf0 by '[] patch'. Original change by elenairina@elenairina:incompatible-guava5:667:citc on 2016/08/24 08:03:39.
Original change ignored the fact that LinkedListMultimap and ConcurrentHashMap
handle the get operation differently. The first one returns an empty collection when
there is nothing associated with a given key in the data structure, while the
latter simply returns null.


Removing some of GUAVA dependencies from junit.runner.junit4 and -.model

Bazel users that are using a different Guava version than the one in the
junitrunner jar are getting an IncompatibleClassChangeError. Rewriting
parts of junitrunner code so it won't depend on Guava anymore.

Continuing progress on issue #1150.

--
MOS_MIGRATED_REVID=131695499