Disable module file pruning for the build of other modules.
Pruning here can lead to a insufficient rebuilds which in turn cause
module-out-of-date errors. The problem surfaces if a module A depends on a
module B, but the headers of module A don't actually use any of B's headers.
Then we would currently not need to rebuild A when B changes although we are
storing a dependency in it. If B changes and we then build something that uses
A (a header of it), we mark A as well as all of its transitive dependencies as
inputs. We still don't need to rebuild A, as none of its inputs have changed,
but we do rebuild B now and then the two modules are out of sync.

For now, just disable pruning of header modules while building the modules. In
the long run, it would be even better to not embed a dependency on B into A,
but that requires signficantly larger changes.

--
MOS_MIGRATED_REVID=132039754