Move bin_dir/genfiles_dir to RuleContext
This is so that the repository name can be included in the root.

To avoid breaking Bazel's CI which tests the HEAD code against the latest
release, none of the Skylark rules in Bazel can be updated to use
ctx.{bin,genfiles}_dir yet, we will have to wait for the 0.3.2 release to
actually update the call sites.  However, this change is to effectively
deprecate ctx.configuration.{bin,genfiles}_dir: it undocuments them and I will
announce the change on the list.

RELNOTES[INC]: bin_dir and genfiles_dir are now properties of ctx, not
configuration. That is, to access the bin or genfiles directory from a
Skylark rule, use ctx.bin_dir or ctx.genfiles_dir (not
ctx.configuration.{bin,genfiles}_dir).  At the moment, you can access
{bin,genfiles}_dir from either, but the ctx.configuration version will
stop working in a future release.

--
MOS_MIGRATED_REVID=132319900