Use inheritance to support site-specific options.
Rename BlazeStartupOptions to StartupOptions and turn the latter into a
class that can be subclassed to implement site-specific options
behavior.  This lends itself to easier customization for the
site-specific versions, especially if there is more than one, and gets
rid of the strange build-time parameterization of the old
blaze_startup_options*.cc files.

Note that this change introduces various TODOs.  This is intentional.
This change is quite complex as it is and I want to focus it on
reshuffling the class structure without actually touching the
internals.  The latter can be done in subsequent, more targetted
changes that are easier to review and understand.

--
MOS_MIGRATED_REVID=132555032