Make environment dependency of actions factor through individual values
With actions depending on the (white-listed part) of the environment
as a whole, even though they are only re-executed if the used parts of
the environment change, each action has to be reconsidered on any change
of the environment. For large dependency graphs, this can be a considerable
amount of effort; therefore add intermediate values for the individual
variables and make actions only depend on those actually used.

--
Change-Id: I283d289da3e0782dc4f9ac084a41425166cfede0
Reviewed-on: https://bazel-review.googlesource.com/#/c/5494
MOS_MIGRATED_REVID=133255911