Add deprecation support to Bazel.
Bazel has always had a deprecation attribute, but until now it has
been a no-op. After this change, Bazel will warn when a target with
the deprecated attribute unset depends on one with the deprecated
attribute set.

Like all other warnings, this warning will only be displayed when it
matches the output filter. It is also bypassed if the two targets are
in the same package.

RELNOTES: The deprecation attribute of all rules now causes warnings
to be printed when other targets depend on a target with that
attribute set.

--
MOS_MIGRATED_REVID=133415232