Add minimum OS, simulator version, and simulator device flags for tvOS and watchOS.
Currently, the minimum watchOS version is always fixed to the SDK version being used, which may not always be desired (for example, building with the 3.0 SDK but support running on 2.2). The watch flags aren't being used yet (nor are the tvOS flags), but will pave the way for us to fix that.

The necessary CROSSTOOL support is in unknown commit.

--
MOS_MIGRATED_REVID=133602832